requirements.txt
asgiref==3.6.0
coverage==7.2.3
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.4.0
iniconfig==2.0.0
invoke==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
